src/requirements.txt
azure-functions
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
openai==1.12.0
pyodbc==5.1.0
pydantic==2.6.1
python-dotenv==1.0.1
